requirements.txt
python-dotenv
pydantic
langchain
langchain-community
langchain-core
streamlit
langchain-openai
langchain-groq
defusedxml
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-huggingface
sentence-transformers
faiss-cpu
pinecone
langchain-pinecone
langchain-chroma
duckduckgo-search
yfinance
matplotlib
autogen-agentchat
autogen-core
autogen-ext
docker
autogen-ext[docker]